move the testing util module to the utils folder
